Kovri: create core Instance/Configuration + rewrite client Instance
- Separates client/core Instance/Configuration implementation
- Introduces better RAII and preliminary API
- Related refactoring + implementing
