+++ src/client/CMakeLists.txt
@@ -1,7 +1,7 @@
 include_directories("../")
 
 set(CLIENT_SRC
-	"instance.cc"
+  "instance.cc"
   "address_book/impl.cc"
   "address_book/storage.cc"
   "api/datagram.cc"
@@ -16,7 +16,7 @@ set(CLIENT_SRC
   "reseed.cc"
   "service.cc"
   "tunnel.cc"
-	"util/config.cc"
+  "util/config.cc"
   "util/http.cc"
   "util/json.cc"
   "util/parse.cc"
